Fix date, weight, OD/OPH view
Contributes to #15:
The date now autopopulates for previous LDRs.
The weight validation is now working correctly with checking for
integers (0 specifically).
Everything should be done for this now.

Fixes #113:
Previously the OD and OPH HCPs were not authorized to view obstetrics
records or LDRs. This change gives them that authority.